requirements.txt
openai==1.1.1
python-telegram-bot==13.7
PyYAML==6.0.1
requests